pagepull_backend/requirements.txt
django
google-generativeai
selenium
beautifulsoup4
groq
python-dotenv
gunicorn
django-cors-headers
requests
lxml